Latest patch
While testing, I discovered an issue with previewing twentyeleven and
the widget "Twenty Eleven Ephemera." I can add the widget just fine,
but making changes to the widget doesn't seem to work. All I get is a
spinner. This is true both with and without 27767-05.patch​ applied,
so when previewing a theme that registers widgets and interacting with
those widgets, the widget customizer doesn't seem to work quite work
right.