Send nonces to Customizer upon successful login
Prevent cheatin' message after re-authenticating in Customizer.
If the user's session expired while in the Customizer, and they
were prompted to re-authenticate inside the Preview, before this
the Customizer would throw up a cheatin message because the nonce
used to get request the preview or to save the settings was tied
to the user's previous session which is no longer valid.

I belive the issue was introduced in 4.0 with the user session tokens
since the nonces are now tied to session tokens as opposed to user IDs,
and thus they change with each re-login.